♻️ update bluetooth
